Settings Activity now creates Calendar Events
Calendar event to prompt the user to take their surveys is created upon
start of application. If the user changes the time or day default, then
the event gets updated. Event recurs every week on the day chosen on
this screen. A reminder is created in the form of an alert on the time
chosen.