Prescriptions
Added scrollview to the manage prescription layout.
Added "Time" to dialog warning
AddPrescription now handles being called from the
ManagePrescriptionActivity. fields will be filled out based on a bundle
passed to it and querying the database. It also deals with being paused
by the camera activity by saving the state.
Added methods to the PrescriptionOpenHelper to access information from
the database.
Provider and ProviderAPI no longer needed.
Notification service should now say what the notification time is for
taking a prescription. But this still needs to be tested. May not be
popping up.
Also need to fix notification Ids